Move hashlib to libsawtooth
Moves the `hashlib` module from the validator to libsawtooth. Also makes
the following changes:

* Adds the `hex` and `openssl` crates to libsawtooth, since they're
  required by the module
* Adds a missing copyright header to `hashlib.rs`
* Removes the unnecessary `use` statements at the beginning of
  `hashlib.rs`
* Removes the `openssl` crate from the validator, since it's no longer
  used directly
