+++ validator/Cargo.toml
@@ -12,7 +12,6 @@ log = { version = "0.4", features = ["std"] }
 libc = ">=0.2.35"
 lmdb-zero = ">=0.4.1"
 cbor-codec = "0.7"
-openssl = "0.10"
 protobuf = "2.0"
 python3-sys = "0.2"
 sawtooth = { path = "../libsawtooth", features = ["validator-internals"] }
